feat/refactor: add multi script, separate files to cluster & single server & test/test:multi scripts
